pyproject.toml
[tool.poetry]
name = "songmam"
version = "1.2.12"
description = "A production interface to facebook messenger api."
authors = ["Nutchanon Ninyawee <me@nutchanon.org>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
starlette = "^0.13.4"
loguru = "^0.5.1"
httpx = "^0.13.3"
furl = "^2.1.0"
cacheout = "^0.11.2"
avajana = "^0.2.2"
autoname = "^0.1.2"
parse = "^1.16.0"
path = "^15.0.0"
moshimoshi = "^0.1.1"

[tool.poetry.dev-dependencies]
mock = "^4.0.2"
responses = "^0.10.15"
fastapi = "^0.58.1"
uvicorn = "^0.11.5"
flask = "^1.1.2"
aiofiles = "^0.5.0"
Jinja2 = "^2.11.2"
pytest = "^5.4.3"
faker = "^4.1.1"
pytest-asyncio = "^0.14.0"
python-decouple = "^3.3"
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
